module_3/requirements.txt
urllib3>=2.0.0
beautifulsoup4>=4.12.0
Flask>=2.3,<4
huggingface_hub>=0.23.0
llama-cpp-python>=0.2.90,<0.3.0
psycopg[binary]>=3.1.0
python-dateutil>=2.8.0
